requirements.txt
torchvision
torch
matplotlib
numpy
pillow
pandas
scikit-learn